chore: rename JestAdapter
